requirements.txt
selenium==4.1.0
webdriver-manager==3.5.4
python-telegram-bot==20.0
aiohttp==3.8.1
aiofiles==22.1.0
requests==2.28.1